fix(ui-plugins): missing dynamic margins since last refactor
