+++ packages/ui-plugins/src/plugins/tailwindcss/colors.ts
@@ -1,0 +1,20 @@
+import { converter } from "culori";
+
+import tokens from "./tokens";
+
+const parse = converter("rgb");
+
+const dynamicColors = () => {
+	const result: { [key: string]: string } = {};
+	Object.entries(tokens.colors).forEach(([name, color]) => {
+		const rgb = parse(color);
+		const variable = `--av-${name}`;
+		const fallbackValue = rgb
+			? `${rgb.r * 255} ${rgb.g * 255} ${rgb.b * 255}`
+			: "0 0 0";
+		result[name] = `var(${variable}, rgb(${fallbackValue} / <alpha-value>))`;
+	});
+	return result;
+};
+
+export default dynamicColors();
